Remove mindplay psr implementation
This breaks testing because it causes tests to not be run with the psr
library meaning incompatabilities with that library won't be detected.

This is a problem for example if you update the requirement to allow
simple-cache 3.0. The tests fill pass but the library will be broken if
actually used with 3.0 because of the interface changes.